Allow visibility of Inherited Methods to toggle
when building polymer elements that extend Polymer.Element it can be hard to find your own methods in the long list of inherited methods. This solution has all inherited methods hidden by default and adds an `iron-doc-hide-bar` to toggle inherited methods.

It would be great if we could have something like this.  Any improvements or suggestions.